# Three-projects-in-one
Frankly, it was the funniest and the most entertaining project so far. I have tried to combine various skills: SQL, XAML, C#-skills, HTML, and CSS.
I created a mini-game, piano, and funny virus. However, I faced a few obstacles such as: how to show SQL results in XAML but always is good to have room to improve. Also, it would be fascinating somehow to record music.